+++ instrumentation/opentelemetry-instrumentation-pyramid/tests/pyramid_base_test.py
@@ -24,6 +24,8 @@ class InstrumentationTest:
         helloid = int(request.matchdict["helloid"])
         if helloid == 500:
             raise exc.HTTPInternalServerError()
+        if helloid == 900:
+            raise NotImplementedError()
         return Response("Hello: " + str(helloid))
 
     def _common_initialization(self, config):